[FE-02] Create a proper Header component
