+++ src/Home/Header/Header.less
@@ -1,0 +1,11 @@
+@import '../../variables';
+
+.Header {}
+
+.title {
+  text-align: center;
+  font-size: @font-size-title-header;
+  background: @gradient;
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+}
