+++ src/Home/Home.less
@@ -3,11 +3,3 @@
 .Home {
   color: @color-light;
 }
-
-.title {
-  text-align: center;
-  font-size: @font-size-title-header;
-  background: @gradient;
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-}
